+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 5.6
 - HTML smaller font size for Contact details title in ViewContact.php
 - HTML add Emergency & Custody button to Contact Info tip message in GetStuList.fnc.php
 - Support option groups (`<optgroup>`) by adding 'group' to $extra in Inputs.php
+- Add Class Search Widget functions in ClassSearchWidget.fnc.php
 
 Changes in 5.5.1
 ----------------